SSCSCI-855: fix fortify issues, hardcoded pwd, SSL (#3813)
