+++ src/cftlib/resources/docker/cft-lib-work-allocation.yml
@@ -79,7 +79,7 @@ services:
         POSTGRES_PORT: 5432
         POSTGRES_NAME: cft_task_db
         POSTGRES_USERNAME: postgres
-        POSTGRES_PASSWORD: admin
+        POSTGRES_PASSWORD: "${POSTGRES_PASSWORD:admin}"
         ALLOWED_JURISDICTIONS: wa,ia,sscs,civil
         ALLOWED_CASE_TYPES: asylum,wacasetype,sscs,civil,benefit
         WA_S2S_AUTHORIZED_SERVICES: ccd,ccd_data,ccd_gw,ccd_ps,iac,wa_task_management_api,xui_webapp,wa_task_monitor,camunda_bpm,wa_workflow_api,wa_case_event_handler
