+++ src/main/resources/config/application-local.properties
@@ -1,5 +1,5 @@
 # DEVELOPMENT CONFIG
-spring.profiles=local
+spring.config.activate.on-profile=local
 
 # LOGGING
 logging.level.root=INFO
